refactor: Extracted out footer component, simplified logic
