+++ packages/wallets/src/components/DerivAppsSection/DerivAppsSection.scss
@@ -23,12 +23,3 @@
         user-select: none;
     }
 }
-
-.wallets-success-modal {
-    &__footer {
-        width: 100%;
-        display: flex;
-        justify-content: center;
-        gap: 1rem;
-    }
-}
