+++ packages/wallets/src/components/DerivAppsSection/DerivAppsSuccessFooter.scss
@@ -1,0 +1,8 @@
+.wallets-deriv-apps-success-modal {
+    &__footer {
+        width: 100%;
+        display: flex;
+        justify-content: center;
+        gap: 1rem;
+    }
+}
